wasn't properly handling toggle when toggle === false
The old version was setting toggle to true, when it should have remained false.  On android devices, this meant that the focus was be set to the input field, and the soft keyboard would popup on 'dragstart' since it is triggered by focus and blur.  

On my maps, this meant that whenever the search feature was visible, if i tried to drag the map, the keyboard would popup for a second before the focus switched back to the map, creating a buggy behavior.