Allowing GeoJSON Point records to be fully searched
GeoJSON datasets are composed of Point, MultiPoint, LineString, MultiLineString, Polygons, MultiPolygons, and GeometryCollections.  When a dataset contains Point or MultiPoint data, the sub-layer in the geoJSON layer is an instance of L.Marker which means that it returns true for the first part of the `if` statement, and skips the second part.
By removing the `else` we are allowing the function to search both the `L.Marker` options as well as the `layer.feature` which L.Markers have when they are part of a geoJSON layer.